+++ package.json
@@ -9,9 +9,11 @@
   "scripts": {
     "build": "parcel build Extensions/browser/manifest.json --no-content-hash --no-source-maps --dist-dir dist --no-cache --detailed-report 0",
     "watch": "parcel watch Extensions/browser/manifest.json --no-source-maps --dist-dir dist --no-cache --detailed-report 0",
-    "lint": "eslint --ext css,ts Extensions/browser",
-    "lint-fix": "run-p 'lint --fix'",
-    "test": "run-p lint build"
+    "lint": "run-p 'lint:*'",
+    "lint:ts": "eslint --ext ts Extensions/browser",
+    "lint:css": "stylelint 'Extensions/browser/*.css'",
+    "lint-fix": "run-p 'lint:* --fix'",
+    "test": "run-p lint:* build"
   },
   "dependencies": {
     "webextension-polyfill": "^0.8.0"
