+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
     "compilerOptions": {
-        "module": "system",
+        "module": "CommonJS",
         "removeComments": true,   
     }
 }
